driver/requirements.txt
pyserial==3.5
py2exe==0.13.0.0
